refactor(api): split REST API handlers
Separate route registration, health, deployment runs, scheduled jobs, lifecycle, projects, stacks, and poll handlers into focused files without changing behavior. Split the corresponding tests along the same boundaries.\n\nCo-authored-by: Copilot <223556219+Copilot@users.noreply.github.com>
