+++ cmd/doco-cd/handler_api_health_test.go
@@ -1,0 +1,71 @@
+package main
+
+import (
+	"net/http"
+	"net/http/httptest"
+	"regexp"
+	"testing"
+
+	"github.com/kimdre/doco-cd/internal/config/app"
+
+	"github.com/kimdre/doco-cd/internal/docker"
+	"github.com/kimdre/doco-cd/internal/logger"
+)
+
+// Make http call to HealthCheckHandler.
+func TestHandlerData_HealthCheckHandler(t *testing.T) {
+	t.Parallel()
+
+	expectedResponse := `{"content":"healthy","job_id":"[a-f0-9-]{36}"}`
+	expectedStatusCode := http.StatusOK
+
+	appConfig, err := app.GetConfig()
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	appConfig.GitCommitStatus = false
+
+	log := logger.New(logger.LevelCritical)
+
+	dockerCli, err := docker.CreateDockerCli(appConfig.DockerQuietDeploy)
+	if err != nil {
+		t.Fatalf("Failed to create docker client: %v", err)
+	}
+
+	t.Cleanup(func() {
+		err = dockerCli.Client().Close()
+		if err != nil {
+			return
+		}
+	})
+
+	h := handlerData{
+		dockerCli:  dockerCli,
+		appConfig:  appConfig,
+		appVersion: app.Version,
+		log:        log,
+	}
+
+	req, err := http.NewRequest("GET", healthPath, nil)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	rr := httptest.NewRecorder()
+	handler := http.HandlerFunc(h.HealthCheckHandler)
+	handler.ServeHTTP(rr, req)
+
+	if status := rr.Code; status != expectedStatusCode {
+		t.Errorf("handler returned wrong status code: got %v want %v", status, expectedStatusCode)
+	}
+
+	regex, err := regexp.Compile(expectedResponse)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	if !regex.MatchString(rr.Body.String()) {
+		t.Errorf("handler returned unexpected body: got %v want %v", rr.Body.String(), expectedResponse)
+	}
+}
